Updated components to track position and HP
Updated board to track the position of the players based on
where in the array their names are.
Added state variable to Game to track each player's HP.
Added paragraphs to the GameBoard to show each players HP.
