+++ src/components/GameBoard/GameBoard.js
@@ -1,35 +1,37 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, Fragment } from 'react'
 // import apiUrl from '../../apiConfig'
 // import axios from 'axios'
 import Cell from '../Cell/Cell'
 
-const GameBoard = ({ user, gameId, startingBoard, msgAlert }) => {
+const GameBoard = ({ startingBoard, p1HP, p2HP }) => {
   const [board, setBoard] = useState([])
   useEffect(() => {
     setBoard(startingBoard)
   }, [startingBoard])
   const handleChange = (cellId, value) => {
-    const newBoard = cleanBoard.map(value => value)
+    const newBoard = board.map(value => value)
     if (value) {
       newBoard[cellId] = Number(value)
     } else {
       newBoard[cellId] = ''
     }
   }
-  const cleanBoard = [...board].map(cell => typeof cell === 'number' ? cell : '')
-  const boardJsx = cleanBoard.map((cell, index) => (
+  const boardJsx = board.map((cell, index) => (
     <Cell
       key = {index}
       id = {index}
       handleChange = {handleChange}
-      cell = {index}
+      cell = {cell}
     />
   ))
   return (
-    <form className = 'board' >
-      {boardJsx}
-    <
-    /form>
+    <Fragment>
+      <form className = 'board' >
+        {boardJsx}
+      </form>
+      <p>Player 1 HP <br/>{p1HP}</p>
+      <p>Player 2 HP <br/>{p2HP}</p>
+    </Fragment>
   )
 }
